python-server/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Environment variables
python-dotenv>=1.0.0

# LangChain core + integrations
langchain>=0.3.25
langchain-core>=0.3.58
langchain-community>=0.3.23
langchain-google-genai>=2.1.4
langchain-pinecone>=0.2.13
langchain-text-splitters>=0.3.8
langsmith>=0.7.14

# Pinecone client
pinecone>=6.0.0

# PDF rendering
pymupdf>=1.27.0

# File uploads
python-multipart>=0.0.12

# Retry / rate-limit handling
tenacity>=8.2.0

# LlamaParse (alternative ingestion pipeline)
llama-cloud>=1.0

# PostgreSQL async driver + ORM (Neon)
psycopg[binary]>=3.2.0
sqlalchemy[asyncio]>=2.0.33
greenlet>=3.0.0

# Auth tokens
PyJWT>=2.10.1
